Cities переделан в функциональный компонент, стейт перенесен в HOC
